refactor code . change import Tree's Images way
