+++ 15/Memrise-App/TreeImages.js
@@ -1,0 +1,15 @@
+import tree0 from "./assets/tree/tree0.png";
+import tree1 from "./assets/tree/tree1.png";
+import tree2 from "./assets/tree/tree2.png";
+import tree3 from "./assets/tree/tree3.png";
+import tree4 from "./assets/tree/tree4.png";
+import tree5 from "./assets/tree/tree5.png";
+
+const TreeImages= [tree0,tree1,tree2,tree3,tree4,tree5];
+
+/**
+ * It exports Images of tree
+ * 
+ */
+
+module.exports = TreeImages;
