+++ 15/Memrise-App/__test__/Courses.test.js
@@ -1,8 +1,8 @@
 import React from "react";
 import renderer from "react-test-renderer";
-import Courses from "../../screens/Courses";
+import Courses from "../screens/Courses";
 
 it("Test courses Screen", () => {
-  const tree = renderer.create(<Courses />).toJSON();
+  const tree = renderer.create(<Courses/>).toJSON();
   expect(tree).toMatchSnapshot();
 });
