+++ 15/Memrise-App/__test__/ListWords.test.js
@@ -1,6 +1,6 @@
 import React from "react";
 import renderer from "react-test-renderer";
-import ListWords from "../../screens/ListWords";
+import ListWords from "../screens/ListWords";
 
 it("Test ListWords Screen", () => {
   const tree = renderer.create(<ListWords/>).toJSON();
