+++ 15/Memrise-App/__test__/WordDetail.test.js
@@ -1,6 +1,6 @@
 import React from "react";
 import renderer from "react-test-renderer";
-import WordDetail from "../../screens/WordDetail";
+import WordDetail from "../screens/WordDetail";
 
 it("Test WordDetail Screen", () => {
   const tree = renderer.create(<WordDetail/>).toJSON();
